feat: capitalisation with italian rules
