+++ src/routes/viewer/Table.tsx
@@ -45,7 +45,7 @@ import { Input } from "@/components/ui/input";
 import MeritTable from "@/utils/types/data/parsed/Ranking/MeritTable";
 import CourseTable from "@/utils/types/data/parsed/Ranking/CourseTable";
 import Store from "@/utils/data/store";
-import { sha256 } from "@/utils/strings";
+import { sha256 } from "@/utils/strings/crypto";
 
 interface TableProps extends React.HTMLAttributes<HTMLTableElement> {
   school: School;
