+++ src/utils/strings.ts
@@ -1,33 +1,5 @@
-import { SALT } from "./constants";
-
-export function capitalizeWords(str: string): string {
-  const words = str.split(/\b/);
-  const capitalizedWords = words.map((word) => {
-    const firstLetter = word.charAt(0).toUpperCase();
-    const restOfWord = word.slice(1).toLowerCase();
-    return firstLetter + restOfWord;
-  });
-  return capitalizedWords.join("");
-}
-
 export function containsOnlyNumbers(input?: string): boolean {
   if (!input) return false;
   return /^[0-9]+$/.test(input);
 }
 
-export async function sha256(input: string, salt = SALT): Promise<string> {
-  // encode as UTF-8
-  const msgBuffer = new TextEncoder().encode(input + salt);
-
-  // hash the input
-  const hashBuffer = await crypto.subtle.digest("SHA-256", msgBuffer);
-
-  // convert ArrayBuffer to Array
-  const hashArray = Array.from(new Uint8Array(hashBuffer));
-
-  // convert bytes to hex string
-  const hashHex = hashArray
-    .map((b) => ("00" + b.toString(16)).slice(-2))
-    .join("");
-  return hashHex;
-}
